+++ gunicorn/test/test_gunicorn.py
@@ -29,10 +29,10 @@ class TestGunicorn(AgentCheckTest):
         """
         Testing Gunicorn check.
         """
-        self.run_check({'instances': [{'proc_name': 'dd-test-gunicorn'}]})
+        self.run_check({'instances': [{'proc_name': 'dd-test-gunicorn', 'tags': ['optional:tag1']}]})
 
-        self.assertMetric("gunicorn.workers", tags=['app:dd-test-gunicorn', 'state:idle'], at_least=0)
-        self.assertMetric("gunicorn.workers", tags=['app:dd-test-gunicorn', 'state:working'], at_least=0)
+        self.assertMetric("gunicorn.workers", tags=['app:dd-test-gunicorn', 'state:idle', 'optional:tag1'], at_least=0)
+        self.assertMetric("gunicorn.workers", tags=['app:dd-test-gunicorn', 'state:working','optional:tag1'], at_least=0)
 
         self.assertServiceCheck("gunicorn.is_running", count=1)