+++ gitlab/CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG - gitlab
 
+1.1.0 / Unreleased
+==================
+
+### Changes
+
+* [FEATURE] Add support for instance level checks in service check.
+
 1.0.0 / 2018-01-10
 ==================