+++ gitlab_runner/CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG - gitlab_runner
 
+1.1.0 / Unreleased
+==================
+
+### Changes
+
+* [FEATURE] Add custom tag support.
+
 1.0.0 / 2018-01-10
 ==================