+++ ntp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG - ntp
 
+1.2.0 / Unreleased
+==================
+
+### Changes
+
+* [FEATURE] adds custom tag support.
+
 1.0.0 / 2017-03-22
 ==================