+++ postgres/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG - postgres
 
+2.1.0 / 2018-03-07
+==================
+### Changes
+
+* [BUGFIX] Adding support for postgres 10. See [#1172][].
+
 2.0.0 / 2018-02-13
 ==================
 ### Changes
@@ -68,5 +74,8 @@
 [#699]: https://github.com/DataDog/integrations-core/issues/699
 [#740]: https://github.com/DataDog/integrations-core/issues/740
 [#776]: https://github.com/DataDog/integrations-core/issues/776
+[#1042]: https://github.com/DataDog/integrations-core/issues/1042
+[#1073]: https://github.com/DataDog/integrations-core/issues/1073
+[#1172]: https://github.com/DataDog/integrations-core/issues/1172
 [@Erouan50]: https://github.com/Erouan50
 [@infothrill]: https://github.com/infothrill